refactor: moved the content of the path_manager file to the path_provider
