+++ path/provider/path_provider.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 import argparse
+from path.validator.path_validator import IPathValidator
 
 
 class IPathProvider(ABC):
@@ -19,5 +20,21 @@ class CliPathProvider(IPathProvider):
 
         args = parser.parse_args()
         path = args.path
+
+        return path
+
+
+class PathManager(IPathProvider):
+    def __init__(
+        self,
+        provider: IPathProvider,
+        validator_manager: IPathValidator,
+    ):
+        self.provider = provider
+        self.validator_manager = validator_manager
+
+    def get(self) -> str:
+        path = self.provider.get()
+        self.validator_manager.validate(path)
 
         return path
